bug_fix: pmm_all_one The size of the index variable in bitmap::set() and bitmap::[] (uin8_t) caused an index overflow each time the bitmap was being indexed with an index greater than 256 causing most reads to show as true even when the address wasn't locked/reserved. The size of index has been changed to 64bits from 8bits to fix this

+++ source/include/kernel/util/bitmap.h
@@ -6,8 +6,8 @@ class bitmap{
 public:
     uint8_t * buffer;
     uint32_t n_bytes;
-    void set(uint8_t index, bool status);
-    bool operator[](uint8_t index);
+    void set(uint64_t index, bool status);
+    bool operator[](uint64_t index);
     bitmap();
     bitmap(uint8_t * buffer, uint32_t n_items);
 };
